split wildcard expnader
