+++ Makefile
@@ -108,9 +108,10 @@ EXPANDER_FILES = $(EXPANDER_DIR)/expander.c \
 				$(EXPANDER_DIR)/expander_variables.c \
 				$(EXPANDER_DIR)/heredoc_expander.c \
 				$(EXPANDER_DIR)/mixed_wildcard_handler.c \
+				$(EXPANDER_DIR)/wildcard_array_utils.c \
+				$(EXPANDER_DIR)/wildcard_expander_utils.c\
 				$(EXPANDER_DIR)/wildcard_expander.c \
 				$(EXPANDER_DIR)/wildcard_replacement.c \
-				$(EXPANDER_DIR)/wildcard_array_utils.c \
 				$(EXPANDER_DIR)/wildcard_filesystem.c \
 
 OBJ_FILES = $(SRC_FILES:$(SRC_DIR)/%.c=$(OBJ_DIR)/%.o)
